+++ src/main/java/wooteco/subway/auth/application/AuthorizationException.java
@@ -3,7 +3,7 @@ package wooteco.subway.auth.application;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
-@ResponseStatus(HttpStatus.UNAUTHORIZED)
+@ResponseStatus(value = HttpStatus.BAD_REQUEST, reason = "이메일과 비밀번호를 확인해주세요.")
 public class AuthorizationException extends RuntimeException {
     public AuthorizationException() {
     }
