+++ src/main/java/wooteco/subway/member/application/DuplicateEmailException.java
@@ -1,4 +1,4 @@
-package wooteco.subway.member.exception;
+package wooteco.subway.member.application;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
